Remove `NUMERIC` cast migration; add it to the original `up.sql` to avoid an unnecessary migration and cognitive load

+++ rust/processor/src/db/postgres/migrations/2025-01-08-171729_emojicoin_arena/up.sql
@@ -142,12 +142,12 @@ CREATE VIEW arena_leaderboard AS
 WITH melee AS (
     SELECT * FROM arena_melee_events ORDER BY melee_id DESC LIMIT 1
 ), price_emojicoin_0 AS (
-    SELECT avg_execution_price_q64 / POW(2,64) AS price FROM swap_events
+    SELECT avg_execution_price_q64 / POW(2,64)::NUMERIC AS price FROM swap_events
     WHERE market_address = (SELECT emojicoin_0_market_address FROM arena_melee_events WHERE melee_id = (SELECT melee_id FROM melee))
     ORDER BY market_nonce DESC
     LIMIT 1
 ), price_emojicoin_1 AS (
-    SELECT avg_execution_price_q64 / POW(2,64) AS price FROM swap_events
+    SELECT avg_execution_price_q64 / POW(2,64)::NUMERIC AS price FROM swap_events
     WHERE market_address = (SELECT emojicoin_1_market_address FROM arena_melee_events WHERE melee_id = (SELECT melee_id FROM melee))
     ORDER BY market_nonce DESC
     LIMIT 1
